requirements.txt
slackclient
bottle
memorize
six
